MNT: Organize Scripts and optimization codes.
