+++ apsuite/optimization/__init__.py
@@ -1,0 +1,4 @@
+from .genetic_algorithm import GA
+from .pso import PSO
+from .scanning import SimpleScan
+from .simulated_annealing import SimulAnneal
