+++ apsuite/optimization/examples.py
